models/vgg16.py
import torch
import torchvision 
import torch.nn as nn 

class VGG16(nn.Module):
    def __init__(self, pretrained_path):
        super(VGG16, self).__init__()
        vgg16 = torchvision.models.vgg16()
        vgg16.load_state_dict(torch.load(pretrained_path))

        self.enc_1 = nn.Sequential(*vgg16.features[:5])
        self.enc_2 = nn.Sequential(*vgg16.features[5:10])
        self.enc_3 = nn.Sequential(*vgg16.features[10:17])

        for i in range(3):
            for param in getattr(self, f'enc_{i+1:d}').parameters():
                param.requires_grad = False 

    def forward(self, image):
        results = [image]
        for i in range(3):
            func = getattr(self, f'enc_{i+1:d}')
            results.append(func(results[-1]))
        return results[1:]

def build_vgg16(args):
    return VGG16(args.pretrained_feat_extractor)